+++ CHANGELOG.md
@@ -12,9 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	  Lofty writes tags. These are best used as global user-configurable options, as most options will
 	  not apply to all files. The defaults are set to be as safe as possible,
 	  see [here](https://docs.rs/lofty/latest/lofty/struct.WriteOptions.html#impl-Default-for-WriteOptions).
-- **ChannelMask** ([PR](https://github.com/Serial-ATA/lofty-rs/pull/370))
+- **ChannelMask** ([PR](https://github.com/Serial-ATA/lofty-rs/pull/371))
   - `BitAnd` and `BitOr` implementations
   - Associated constants for common channels, ex. `ChannelMask::FRONT_LEFT`
+
+### Changed
+- **Properties**: `FileProperties` and `ChannelMask` have been moved from the root to the new `lofty::properties`
+                   module ([PR](https://github.com/Serial-ATA/lofty-rs/pull/372))
 
 ### Fixed
 - **Vorbis**: Fix panic when reading properties of zero-length files ([issue](https://github.com/Serial-ATA/lofty-rs/issues/342)) ([PR](https://github.com/Serial-ATA/lofty-rs/pull/365))